move to file-check instead and add generic packager
