+++ scripts/generic/generic-packager.ts
@@ -131,6 +131,13 @@ const packager = (font: Font, rebuildFlag: boolean): void => {
     });
   }
 
+  // Write file-list.json
+  const fileList: string[] = [];
+  fs.readdirSync(`${fontDir}/files`).forEach(file => {
+    fileList.push(`fonts/${type}/${fontId}/${file}`);
+  });
+  jsonfile.writeFileSync(`${fontDir}/files/file-list.json`, fileList);
+
   // Write README.md
   const packageReadme = readme({
     fontId,
