+++ scripts/utils/file-check.ts
